Makefile: Build directory creation improvement
To avoid printed error messages the stderr was previously redirected.
For Windows this needed to be redirected to NUL, while for Linux it
had to be redirected to /dev/null. With .sh scripts on Windows the two
variants could get mixed up, which resulted in build failures or visible
mkdir error messages.

To prevent both from happening, the error messages get stored in a
temporary variable to avoid them going out to the log.