Reworked bootloader reentry mechanism
The Landungsbruecke bootloader checks for a magic value to determine,
whether to start the application or to switch into bootloader mode.

For some ICs the default value of the Driver Enable pin has to be 0. For
most ICs it is 1, so the bootloader used to switch the pin on. Now we
need to pass the information, which default value to use, to the
bootloader. For that the data passed to the bootloader has been
extended. The previous extension relied on implicit memory layout done
by the compiler and linker. This broke backwards compatibility. To avoid
this issue, a struct is now used to explicitly define the memory layout.