Updated the StepDir stallGuard function
The stall function of StepDir handles stall events by stopping the
motion. This is done by calling StepDir_stop() and setting a stall-type
stop halting condition.

The function used a static variable to store the last stall state and
subsequently call the stop function only the first time the stall
condition gets reported. This makes it unsafe to use with multiple
channels - a feature built into the function with the channel parameter.

Additionally, the function is part of periodic job, which is documented
as safe to call from a low priority interrupt. This safety stems from
the fact that a stall persists for multiple calls of this function.
Theoretically a write to the haltingCondition variable in an interrupt
can be overwritten by the main loop. In practise the interrupt can just
rewrite the field the next iteration. This requires that no mechanism
that ensures only one call to stop() is present.