StepDir: Moved the stallguard velocity check to periodicJob()
The StepDir interrupt needs to be handled as quickly as possible. In a
properly configured system the stall velocity threshold should be
overstepped with enough margin of error before a stall can occur. As a
result the velocity threshold check does not need to be part of the
realtime loop and can be moved to the periodic job.