Fixed race condition in StepDir_deInit() on Landungsbruecke
When calling StepDir_deInit() an interrupt could be generated just as
the clock gating of the FTM module is getting disabled. As a result the
interrupt attempts to access the now disabled FTM module and receives a
bus fault.

This is mitigated by disabling the FTM module and it's interrupts before
disabling the clock gating.