Fixed error in StepDir_deInit()
The previous fix to a racecondition between the StepDir interrupt and
the deInit() function did assume wrongly, that StepDir would be
initialized whenever called. As a result, the function accessed the FTM1
registers - which turned out to be fatal if StepDir and consequently
clock gating of the FTM1 module was not initialized.