Fixed potential race condition in StepDir_setPins()
When changing a pin to a dummy pin the interrupt could potentially
attempt to use that dummy pin, resulting in a crash if the dummy pin was
declared lazily.

Dummy pins are defined as pins with .bitWeight == 0
The .setBitRegister and .resetBitRegister members may or may not be
defined. If theyre not defined using the pin will result in a NULL
pointer dereference.