Added special case for TMC43XX + TMC2130 combination
The TMC2130 uses the DIO12 pin for the ENCA_DCIN_CFG5 pin. The reset
value is LOW. When running iwth a TMC43XX that pin is not connected and
DIO12 is used by the TMC43XX as STOP_L pin with a reset value of HIGH.

This user function sets the pin back to HIGH after the initialization.