TMC4671: Fixed config status default being set to busy
The TMC4671 does not use the configuration mechanism that the other
chips use. The status of the config mechanism is therefore unused.

The old default value of the status was busy, which resulted in rapid
blinking of the heartbeat LED. This has been changed to a default status
of ready, resulting in the usual slow heartbeat LED.