TMC7300: Removed the Standby error entry from checkErrors()
The checkErrors function is used to collect various errors and warn the
user about them in the TMCL IDE. Being in the standby state is not an
error but rather a status. However the currently implemented mechanisms
does not include a centralized status information collection.

This change also stops the standby turning on the error LED. While this
is a nice indicator, it does not fit well into the convention of an
active error LED indicating that something went wrong.