StepDir: Inlined pin check during interrupt
The current HAL implementation wastes multiple cycles by being
implemented with pointers. In this case this makes a crucial difference,
so for now we simply inline the pin logic. This saves around 20 cycles
in the StepDir interrupt on the Landungsbruecke.

Long-term the HAL should provide functions directly instead of through
pointer indirection.