Timer: Implemented optional external callback on timer overflow interrupt.
This callback can then be used in eval implementations to do custom
timer-based stuff.
The timer implementation currently only uses FTM0 on Landungsbruecke.
For this to work, and as interrupt implementations can only be defined once,
all further multiplexing has to be done with this callback mechanism.
Thus, the BLDC implementation has been modified such that it utilizes
the central FTM0 interrupt in Timer.c forwarded via callback (while
being completely independend in the rest of the FTM0 configuration).