TMC2660: Fixed multiple APs not reading correctly
This fix is not a clean one - the tmc2660_readInt function uses the
following address space: read registers 0-2, write registers 8-15.
The macros used for reading back the write registers were incorrectly
using the address space of the IC (0-7) instead. Due to the intricacies
of the IC<->address space mapping, the issue has only been fixed with a
simple workaround for now.