TMC7300: Changed consistencyCheck interval to one second
The error check is mostly a sanity check in case of chip brownouts,
It does not need to run often. Running it frequently also drastically
lowers the amount of TMCL cmd/s you can have in the TMCL-IDE.

Finally, the 0 check interval caused a "always true" warning about the
timing if() check.