Added support for git information in GET_INFO
This is disabled by default and must be enabled by passing
ENABLE_GIT_VERSION_INFO=1 to make.
When enabled, it requires git and python to be installed.
By default, make expects both a "git" and a "python" binary to be
present in the PATH. If this is not the case, the locations and names of
these binaries can instead be supplied by setting the GIT and PYTHON
variables accordingly.